Split deleteListing logic to separate function
