+++ src/app/api/listing/[listing_id]/route.ts
@@ -1,7 +1,6 @@
 import { NextResponse } from "next/server";
-import {db} from "@/lib/firebase/config";
-import { User, Listing, newListing } from "@/lib/firebase/firestore/types";
-import { getDoc, doc, updateDoc, arrayRemove, deleteDoc } from "firebase/firestore";
+import { newListing } from "@/lib/firebase/firestore/types";
+import deleteListing from "@/lib/firebase/firestore/listing/deleteListing";
 
 /*
  * Get a Listing by id
@@ -81,57 +80,9 @@ export async function DELETE(
     // get user id from req body
     const {user_id} = await req.json();
 
-    const listingRef = doc(db, 'listings', listing_id);
-    const listingSnapshot = await getDoc(listingRef);
+    const ret_id: string = await deleteListing(listing_id, user_id);
 
-    if (!listingSnapshot.exists()) {
-      throw new Error("Listing not found");
-    }
-
-    const listingData = listingSnapshot.data() as Listing;
-    if (!listingData) {
-      throw new Error("Listing data invalid");
-    }
-
-    // extract relevant fields
-    const {owner, potential_buyers} = listingData;
-
-    if (user_id === owner) {
-      const ownerRef = doc(db, 'users', owner);
-      const ownerSnapshot = await getDoc(ownerRef);
-      if (!ownerSnapshot.exists()) {
-        // do not throw an error from API call, but log problem
-        console.error(`owner ${owner} not found`);
-      }
-      const ownerData = ownerSnapshot.data() as User;
-      if (ownerData) {
-        await updateDoc(ownerRef, { active_listings: arrayRemove(listing_id) });
-      } else {
-        // do not throw an error from API call, but log problem
-        console.error(`owner ${owner} data invalid`);
-      }
-      await Promise.all(potential_buyers.map(async (buyer) => {
-        const buyerRef = doc(db, 'users', buyer);
-        const buyerSnapshot = await getDoc(buyerRef);
-        if (!buyerSnapshot.exists()) {
-          // do not throw an error from API call, but log problem
-          console.error(`potential buyer ${buyer} not found`);
-        }
-        const buyerData = buyerSnapshot.data() as User;
-        if (buyerData) {
-          await updateDoc(buyerRef, { interested_listings: arrayRemove(listing_id) });
-        } else {
-          // do not throw an error from API call, but log problem
-          console.error(`owner ${owner} data invalid`);
-        }
-      }));
-    } else {
-      throw new Error("Unauthorized user");
-    }
-
-    await deleteDoc(listingRef);
-
-    return NextResponse.json({ data: { listing_id: listing_id }, error: null });
+    return NextResponse.json({ data: { listing_id: ret_id }, error: null });
   } catch (e: unknown) {
     if (e instanceof Error) {
       return NextResponse.json({ data: null, error: e.message });
